+++ packages/sewing-kit-koa/src/index.ts
@@ -1,2 +1,2 @@
 export {default as Assets} from './assets';
-export {default as middleware, State} from './middleware';
+export {default as middleware, getAssets, setAssets} from './middleware';